gpstate: reimplement segment status using pg_isready
pg_isready gives us basic information: is the segment up or down, and if
it's up, is it a primary or a mirror? It can also tell whether a segment
is starting or stopping, but it can't tell us *which* of those two
states it's in -- only that the segment is alive but rejecting
connections.

Co-authored-by: Shoaib Lari <slari@pivotal.io>
Co-authored-by: Mark Sliva <msliva@pivotal.io>