behave: fix "user kills * process" steps
These steps relied on has_process_eventually_stopped() to tell them when
a process of a given PID had finally exited. Unfortunately that function
doesn't take a PID -- it takes a process name and passes that to pgrep.

Replace the implementation here with one that supports PIDs. (And bring
the default timeout down from 2 minutes; there's no reason a process
kill should take that long.)