gpinitsystem: Return 0 on warnings
Previously gpinitsystem returned 1 in case of warngings (2 in case of
errors). We have changed gpinitsystem to return 0 for warnings so that other
tools using gpinitsystem (such as gpupgrade) will not fail when gpinitsystem
generates warnings.

Also, gpinitsystem was exiting with non-zero return code if certain errors or
warnings had happened on any previous run of gpinitsystem that day. This has
also been fixed in this commit

Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>
Co-authored-by: Shoaib Lari <slari@pivotal.io>